% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2distance.R
\name{l2dist}
\alias{l2dist}
\title{Function that computes the distance between the two functions estimated by the samples}
\usage{
l2dist(sample1, sample2, samp = FALSE, rancor = 0.05, gridsize = NA)
}
\arguments{
\item{sample1}{a vector of probabilties for the support points}

\item{sample2}{a vector of probabilties for the support points}

\item{samp}{noclue}

\item{rancor}{noclue}

\item{gridsize}{noclue}
}
\value{
dist L2 - distance between the two functions estimated by the samples
}
\description{
Function that computes the distance between the two functions estimated by the samples
}
\examples{
#l2dist(s)
}
