% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{med.list}
\alias{med.list}
\title{med.list}
\usage{
med.list(beta_list)
}
\arguments{
\item{beta_list}{a two dimensional vector of which the ordered median is of interest}
}
\value{
vector of increasing medians of
}
\description{
Computes the increasing order of items based on median. This is the basis for ordering a list according to the range of vectors in the list.
}
\examples{
#med.list(list(x=runif(100,min=1,max=2),y=runif(100,min=1,max=2)))
}
