% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{ran.list}
\alias{ran.list}
\title{ran.list}
\usage{
ran.list(beta_list)
}
\arguments{
\item{beta_list}{a two dimensional vector of which the ordered range is of interest}
}
\value{
vector of ordered ranges
}
\description{
Computes the vector of indices according to decreasing order of the ranges within the vector. This is the basis for ordering a list according to the range of vectors in the list.
}
\examples{
#ran.list(list(x=runif(100,min=1,max=2),y=runif(100,min=1,max=2)))
}
