% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{ran}
\alias{ran}
\title{Derives the range of a vector}
\usage{
ran(x)
}
\arguments{
\item{x}{vector with unknown range}
}
\value{
the range of the vector
}
\description{
Derives the range of a vector
}
\examples{
custom_vector = c(5,2,7,9,4)
ran(custom_vector)

}
