% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yin.yang.sampling.R
\name{yin.yang.sampling}
\alias{yin.yang.sampling}
\title{A method that does stuff}
\usage{
yin.yang.sampling(
  sample_list = NULL,
  prior_list = NULL,
  algo = "multinom",
  method = "sequential",
  prob_old = "heuristic",
  burn.in = 2500,
  bw.adj = 1,
  rancor = 1,
  neps = 500,
  check = FALSE,
  nsam = NULL
)
}
\arguments{
\item{sample_list}{parameter}

\item{prior_list}{parameter}

\item{algo}{parameter}

\item{method}{parameter}

\item{prob_old}{parameter}

\item{burn.in}{parameter}

\item{bw.adj}{parameter}

\item{rancor}{parameter}

\item{neps}{parameter}

\item{check}{parameter}

\item{nsam}{parameter}
}
\value{

}
\description{
A method that does stuff
}
\examples{

}
