% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2distance.R
\name{l2dist}
\alias{l2dist}
\title{L2Distance}
\usage{
l2dist(sample1, sample2, samp = FALSE, rancor = 0.05, gridsize = NA)
}
\arguments{
\item{sample1}{a vector of observations}

\item{sample2}{a vector of observations}

\item{samp}{indicator whether a random subsample should be sampled or the full original observations taken for calculations}

\item{rancor}{value of how much larger the support is the kernel density estimator is chosen compared beyond the minimum and maximum value}

\item{gridsize}{factor of the grid size chosen for kernel density estimator}
}
\value{
dist L2 - distance between the two functions estimated by the samples
}
\description{
Function that computes the L2 distance between the two functions estimated by the samples
}
\examples{
#x<-runif(100);y<-runif(100)
#l2dist(x,y)
}
