% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{ran.list}
\alias{ran.list}
\title{Computes the vector of the highest ranges within the vector and returns it as a vector of decreasing size}
\usage{
ran.list(beta_list)
}
\arguments{
\item{beta_list}{a two dimensional vector of which the ordered range is of intersest}
}
\value{
vector of ordered ranges
}
\description{
Computes the vector of the highest ranges within the vector and returns it as a vector of decreasing size
}
\examples{

}
